fix: enhance error handling in `GetDebtCountByService` and update `ServiceDebtReport.Count` type to `int64`
- Add validation for `countVal` to handle invalid types and return standardized HTTP error.
- Update `get_debt_count_by_service_test` to use `int64` for accurate type matching.

+++ neo4jrepositories/reportrepository/get_debt_count_by_service_test.go
@@ -134,7 +134,7 @@ func TestNeo4jReportRepository_GetDebtCountByService_FilterAndExclude(t *testing
 	}
 
 	// Convert to map for easy assertions
-	got := make(map[string]int)
+	got := make(map[string]int64)
 	names := make(map[string]string)
 	for _, r := range report {
 		got[r.Id] = r.Count
